user-server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
psycopg2-binary
pydantic-settings
httpx
docker
RestrictedPython
eth-account
eth-abi
eth-utils
APScheduler
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
openai
anthropic
google-generativeai
